Add auto mapper

+++ Source/Domain/Morsley.UK.YearPlanner.Users.Application/Handlers/DeleteUserCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using MediatR;
 using Morsley.UK.YearPlanner.Users.Application.Commands;
 using Morsley.UK.YearPlanner.Users.Domain.Interfaces;
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -12,11 +13,13 @@ namespace Morsley.UK.YearPlanner.Users.Application.Handlers
 
         public DeleteUserCommandHandler(IUnitOfWork unitOfWork)
         {
-            _unitOfWork = unitOfWork;
+            _unitOfWork = unitOfWork ?? throw new ArgumentNullException(nameof(unitOfWork));
         }
 
         public async Task<Unit> Handle(DeleteUserCommand command, CancellationToken ct)
         {
+            if (command == null) throw new ArgumentNullException(nameof(command));
+
             _unitOfWork.UserRepository.Delete(command.Id);
             var numberOfRowsAffected = await _unitOfWork.CompleteAsync();
             // ToDo --> Log!
